SQL문 수정 스토어 리스트 카테고리별로 뿌리기

+++ src/main/java/com/koreait/cgvproject/dto/GiftDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class GiftDTO{
-    private Integer gcode;
+    private Long gcode;
     private String category;
     private String title;
     private String gname;
