+++ src/main/java/com/koreait/cgvproject/entity/Gift.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 
@@ -19,7 +20,7 @@ public class Gift{
 
     @Id
     @GeneratedValue( strategy = GenerationType.SEQUENCE, generator = "seq_gift")
-    private Integer gcode;
+    private Long gcode;
     private String category;
     private String title;
     private String gname;
